+++ common/src/main/java/com/julun/huanque/common/bean/forms/Form.kt
@@ -75,7 +75,7 @@ class SettingForm(
 class LiveRemindForm(
     var offset: Int? = null,
     var pushOpen: String? = null,
-    var programId: Int? = null
+    var programId: Long? = null
 )
 
 class FindNewsForm {
