+++ core/src/main/java/com/julun/huanque/core/viewmodel/LiveSquareViewModel.kt
@@ -54,7 +54,7 @@ class LiveSquareViewModel : BaseViewModel() {
     }
 
     private var offsetHot = 0
-    fun requestHotList(queryType: QueryType, programId: Int) {
+    fun requestHotList(queryType: QueryType, programId: Long) {
 
         viewModelScope.launch {
             if (queryType == QueryType.REFRESH) {
